Use natural keys for permissions
The current data was relying on the lack of hash randomisation. With
python -R it was already failing. Any change in content types (like
renaming a model) would cause this to break in very misterious ways.
For example custom permissions for glossaries would be missing since
they were associatd with the wrong content type, but the fixture still
seemed to work.