Restore limited ability to add some models
We want to limit the ability to add arbitrary translations, definitions
and external resources using the admin. The generic form is expensive
and unwieldy (you have to select from all concepts). If the UI provides
a URL with the concept specified, it is easier to use, and not expensive.
A superuser can still use the full form.