Gobbledygook to convince IE to work a bit
May this be the last time in my life that I have to research IE bugs and
implement workarounds for it. The HTML5 doctype was supposed to be good
enough, but this time it seems to be necessary when the site might be on
your company intranet. Or something else. Or not.