Simplify export form
It was way to complicated, and too hard to understand or explain. Now
Django handles all form validation automatically, and I hope it is
simpler to undertand. This should be functionaly equivalent. We should
probably rather let the user choose from a list of all the administrative
statusses in the database, but then enforcing a relative order will be
harder. Maybe we don't need to?