Own forms for permissions in GlossaryAdmin
This provides (in my opinion) a more usable GUI for permissions handling
on Glossaries. It only shows the terminator roles, and doesn't require
knowing about the Guardian action button at the top.

Some care is taken here to implement a lot of the queries by hand, since
the Guardian shortcuts are inefficient with large numbers.